third_party/blink/web_tests/external/wpt/css/css-page/page-orientation.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] contents for selector ['named6']
  assert_equals: unexpected @page contents expected "page-orientation: rotate-right;" but got "page-orientation: initial;"
[FAIL] contents for selector ['h5']
  assert_equals: unexpected style rule contents expected "display: block;" but got "page-orientation: rotate-right; display: block;"
Harness: the test ran to completion.
